Add files via upload
MALWARE DETECTING SYSTEM USING MACHINE LEARNING
Malware continues to be one of the biggest hazards to online safety, and cybersecurity threats 
are always changing.  Traditional signature-based malware detection methods are not keeping 
up with the increasing volume of data and sophistication of attacks.  The goal of this project 
is to create a malware detection system that uses machine learning (ML) techniques to 
determine if a file is harmful or not.  The desire to improve detection capabilities through 
intelligent systems that can learn from data and adjust to new virus types is what inspired this 
study.  Instead of using static signatures, machine learning provides a scalable and dynamic 
method for identifying trends in file activity and detecting malware based on attributes. 
This topic is important in both academia and industry because automated and intelligent 
malware detection is critical for sustaining cybersecurity at the personal, corporate, and 
national levels.  The project will use publicly available malware datasets and supervised 
learning approaches with Python-based ML modules.